pyproject.toml
[tool.poetry]
name = "django_react_admin"
version = "0.1.0"
description = ""
authors = ["ph <robotnaoborot@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.10"
django_filters = "^0.2.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
